Switch to `#!/usr/bin/env sh`, use in ./git_config_kaitai_bot too
The `#!/usr/bin/env sh` shebang is said to be more portable than
`#!/bin/sh`.

I have also added the `-u` shell option (equivalent to `-o nounset`). It
should prevent the usage of undefined variables. I think it should be
supported by a POSIX shell (it's listed here:
<https://pubs.opengroup.org/onlinepubs/9699919799/utilities/V3_chap02.html#set>
and I checked that at least the `dash` supports it), but I'm not sure.